[Radix Cache] Separate Python TreeCore inspector
