+++ test/registered/unit/mem_cache/test_tree_core_registry.py
@@ -18,6 +18,9 @@ from sglang.srt.mem_cache.unified_cache.tree_core_registry import (
     registered_tree_core_backends,
 )
 from sglang.srt.mem_cache.unified_cache.unified_tree_core import UnifiedTreeCore
+from sglang.srt.mem_cache.unified_cache.unified_tree_core_inspection_interface import (
+    UnifiedTreeCoreInspectionInterface,
+)
 from sglang.srt.mem_cache.unified_radix_cache import UnifiedRadixCache
 from sglang.test.ci.ci_register import register_cpu_ci
 from sglang.test.test_utils import CustomTestCase
@@ -98,6 +101,7 @@ class TreeCoreRegistryTest(CustomTestCase):
             components={ComponentType.FULL: component},
         )
         self.assertIsInstance(core, UnifiedTreeCore)
+        self.assertNotIsInstance(core, UnifiedTreeCoreInspectionInterface)
         self.assertIs(component.tree_core, core)
 
     def test_unknown_backend_raises_naming_the_known_backends(self):
